chromium: Support building with NaCl
This commits adds a new 'config.chromium.enableNaCl' option that makes
it possible to build chromium with native client support enabled.

The NaCl and PNaCl toolchains are required to build chromium with NaCl
support. These are not part of the downloaded tarball, and must be
built as a separate derivation.

'common.nix' linkes the toolchains into place if 'enableNaCl' is true.
It's also been changed to do stripping manually, to avoid stripping the
nacl_irt_*.nexe file, as this results in a segfault when trying to run
NaCl apps.